perf: better yeild control during pauses
in stead of using mq.delay for the pauses we are going to pause the nav coroutine and yield control to main until the time has passed.

This allows our interactions with the main gui to be more responsive during long pauses at a waypoint.

Also added an exit button to the menubar. the X on the titlebar will only close the window /mypaths quit will exit but the new exit button below the close button will do the same as well now.